feat: add scheduled games API endpoints and UI tests
- Implement scheduled games retrieval and creation endpoints
- Add tests for scheduled games API functionality
- Move create game UI tests to the appropriate test suite
- Enhance test coverage for scheduled games UI interactions
